# docs/api/backend-openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/token": {
      "post": {
        "description": "Enabled only when ADMIN_API_KEY is configured; returns 503 otherwise. Rate limited per client address (AUTH_RATE_LIMIT_MAX per AUTH_RATE_LIMIT_TTL_SECONDS).",
        "operationId": "AuthController_issueToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many token-issuance attempts from this client.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Exchange an administrative API key for a JWT",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticatedUserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Return the authenticated administrative principal",
        "tags": [
          "auth"
        ]
      }
    },
    "/policies": {
      "post": {
        "description": "Requires an authenticated principal: creation consumes the provider's coverage reserve and is submitted on-chain from Stage 06 onward.",
        "operationId": "PoliciesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePolicyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePolicyResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "Live on-chain execution is wired in Stage 06.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a parametric rainfall policy",
        "tags": [
          "policies"
        ]
      },
      "get": {
        "operationId": "PoliciesController_list",
        "parameters": [
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Zero-based index of the first item to return.",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Maximum number of items to return (capped at 100).",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "insured",
            "required": false,
            "in": "query",
            "description": "Filter policies by the insured account address. Case-insensitive: the value is normalized before it is matched against on-chain data.",
            "schema": {
              "example": "0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyListResponseDto"
                }
              }
            }
          },
          "501": {
            "description": "Live on-chain reads are wired in Stage 06.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "summary": "List policies with pagination",
        "tags": [
          "policies"
        ]
      }
    },
    "/policies/{address}": {
      "get": {
        "operationId": "PoliciesController_getByAddress",
        "parameters": [
          {
            "name": "address",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyResponseDto"
                }
              }
            }
          },
          "501": {
            "description": "Live on-chain reads are wired in Stage 06.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Get a policy by contract address",
        "tags": [
          "policies"
        ]
      }
    },
...
  },
  "info": {
    "title": "ClimateChain API",
    "description": "Parametric climate micro-insurance backend API. Coordinates policy lifecycle, premium pricing, and on-chain integration.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "IssueTokenDto": {
        "type": "object",
        "properties": {
          "apiKey": {
            "type": "string",
            "description": "Administrative API key (matched against ADMIN_API_KEY) exchanged for a short-lived JWT."
          }
        },
        "required": [
          "apiKey"
        ]
      },
      "TokenResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "Signed JWT access token."
          },
          "tokenType": {
            "type": "string",
            "example": "Bearer",
            "description": "Token scheme."
          },
          "expiresIn": {
            "type": "string",
            "example": "1d",
            "description": "Token lifetime as configured by JWT_EXPIRES_IN."
          }
        },
        "required": [
          "accessToken",
          "tokenType",
          "expiresIn"
        ]
      },
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400,
            "description": "HTTP status code."
          },
          "error": {
            "type": "string",
            "example": "Bad Request",
            "description": "Short, machine-stable error label."
          },
          "message": {
            "description": "Human-readable message or list of validation messages explaining the failure.",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": [
              "coverageEth must be a positive number"
            ]
          },
          "method": {
            "type": "string",
            "example": "POST",
            "description": "HTTP method of the request."
          },
          "path": {
            "type": "string",
            "example": "/policies",
            "description": "Request path that produced the error."
          },
          "timestamp": {
            "type": "string",
            "example": "2026-04-07T12:00:00.000Z",
            "description": "ISO-8601 timestamp when the error was produced."
          },
          "requestId": {
            "type": "string",
            "nullable": true,
            "description": "Correlation id matching the `x-request-id` response header.",
            "example": "f0c1d2e3-4567-8901-abcd-ef0123456789"
          }
        },
        "required": [
          "statusCode",
          "error",
          "message",
          "method",
          "path",
          "timestamp"
        ]
      },
      "AuthenticatedUserResponse": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "example": "admin",
            "description": "Principal identifier."
          },
          "roles": {
            "example": [
              "admin"
            ],
            "description": "Roles granted to the principal.",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "userId",
          "roles"
        ]
      },
      "CreatePolicyDto": {
        "type": "object",
        "properties": {
          "coverageEth": {
            "type": "string",
            "example": "1.0",
            "description": "Coverage (payout) amount in ETH."
          },
          "premiumEth": {
            "type": "string",
            "example": "0.05",
            "description": "Premium to fund in ETH (sent as transaction value). Must be at least 1% of coverage on-chain."
          },
          "rainfallThresholdMm": {
            "type": "number",
            "example": 50,
            "minimum": 1,
            "description": "Rainfall threshold in millimeters that triggers a payout."
          },
          "durationDays": {
            "type": "number",
            "example": 30,
            "minimum": 1,
            "maximum": 365,
            "description": "Coverage window length in days."
          },
          "region": {
            "type": "string",
            "example": "Valencia",
            "maxLength": 31,
            "description": "Human-readable region code. Encoded to bytes32 on-chain when provided; must be non-empty (an empty code maps to the on-chain zero region)."
          },
          "requestedStartTimestamp": {
            "type": "number",
            "example": 1767225600,
            "description": "Requested coverage start as a Unix timestamp (seconds). Defaults to a near-future start on-chain when omitted. Requires `region` to be set, since an explicit start is only honored on the on-chain metadata path. Must be at least 180 seconds ahead so it still satisfies the on-chain lead time once the transaction is mined."
          }
        },
        "required": [
          "coverageEth",
          "premiumEth",
          "rainfallThresholdMm",
          "durationDays"
        ]
      },
      "PolicyStatus": {
        "type": "string",
        "enum": [
          "created",
          "active",
          "triggered",
          "paid_out",
          "expired"
        ]
      },
      "CreatePolicyResponseDto": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "description": "Created policy contract address."
          },
          "transactionHash": {
            "type": "string",
            "description": "Transaction hash of the creation transaction."
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PolicyStatus"
              }
            ]
          }
        },
        "required": [
          "address",
          "transactionHash",
          "status"
        ]
      },
      "PolicySettlementType": {
        "type": "string",
        "enum": [
          "none",
          "payout",
          "expiry"
        ],
        "description": "How the provider settled its reserve for this policy. Distinct from `status`, which tracks the policy's own lifecycle."
      },
      "PolicyResponseDto": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "description": "Policy contract address."
          },
          "insured": {
            "type": "string",
            "description": "Insured account address."
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PolicyStatus"
              }
            ]
          },
          "coverageWei": {
            "type": "string",
            "description": "Coverage amount in wei.",
            "example": "1000000000000000000"
          },
          "premiumWei": {
            "type": "string",
            "description": "Premium amount in wei.",
            "example": "50000000000000000"
          },
          "rainfallThresholdMm": {
            "type": "string",
            "description": "Rainfall threshold in millimeters.",
            "example": "50"
          },
          "latestRainfallMm": {
            "type": "string",
            "description": "Latest observed rainfall in millimeters.",
            "example": "0"
          },
          "conditionMet": {
            "type": "boolean",
            "description": "Whether the payout condition has been met."
          },
          "pendingPayoutWei": {
            "type": "string",
            "description": "Coverage amount currently claimable by the insured, in wei. The contract settles payouts pull-style: once a payout executes, the insured must call the policy's claim entry point to receive the funds. A non-zero value here means money is waiting to be claimed.",
            "example": "0"
          },
          "lastOracleUpdateTimestamp": {
            "type": "number",
            "description": "Unix timestamp (seconds) of the last oracle weather update, or 0 if the policy has never been updated. Lets consumers detect stale data.",
            "example": 0
          },
          "oracle": {
            "type": "string",
            "description": "Weather oracle adapter bound to this policy at deployment."
          },
          "regionCode": {
            "type": "string",
            "description": "Region code as a bytes32 hex string.",
            "example": "0x0000000000000000000000000000000000000000000000000000000000000000"
          },
          "startTimestamp": {
            "type": "number",
            "description": "Coverage start as a Unix timestamp (seconds).",
            "example": 1767225600
          },
          "endTimestamp": {
            "type": "number",
            "description": "Coverage end as a Unix timestamp (seconds).",
            "example": 1769817600
          },
          "paidOut": {
            "type": "boolean",
            "description": "Whether the coverage payout has been settled."
          },
          "settlementType": {
            "description": "How the provider settled its reserve for this policy. Distinct from `status`, which tracks the policy's own lifecycle.",
            "allOf": [
              {
                "$ref": "#/components/schemas/PolicySettlementType"
              }
            ]
          },
          "settledAt": {
            "type": "number",
            "nullable": true,
            "description": "Unix timestamp (seconds) when the provider recorded settlement, or null while the policy is unsettled.",
            "example": null
          }
        },
        "required": [
          "address",
          "insured",
          "status",
          "coverageWei",
          "premiumWei",
          "rainfallThresholdMm",
          "latestRainfallMm",
          "conditionMet",
          "pendingPayoutWei",
          "lastOracleUpdateTimestamp",
          "oracle",
          "regionCode",
          "startTimestamp",
          "endTimestamp",
          "paidOut",
          "settlementType"
        ]
      },
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "description": "Total number of items available.",
            "example": 42
          },
          "offset": {
            "type": "number",
            "description": "Offset applied to this page.",
            "example": 0
          },
          "limit": {
            "type": "number",
            "description": "Limit applied to this page.",
            "example": 20
          },
          "count": {
            "type": "number",
            "description": "Number of items in this page.",
            "example": 20
          }
        },
        "required": [
          "total",
          "offset",
          "limit",
          "count"
        ]
      },
      "PolicyListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PolicyResponseDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
    }
  }
}